HV-1155 Move SetClassLoader privileged action to a shared package
